refactor: +env variable for resetting table +file structure refactor
