+++ gambitto-backend/.env
@@ -4,3 +4,5 @@ DB_USER=postgres
 DB_PASSWORD=root
 DB_HOST=localhost
 DB_PORT=5432
+
+RESET_TABLE=false
